requirements.txt
beautifulsoup4==4.11.1
requests==2.26.0
python-dotenv==0.19.0
selenium==4.6.0
